+++ mapentity/static/mapentity/mapentity.js
@@ -20,9 +20,6 @@ $(document).ready(function (e) {
     setTimeout(fillmax, 0);
     $(window).resize(fillmax);
 
-    // Chosen-ify elements
-    $(".chzn-select").chosen();
-
     // Top-navigation tabs
     MapEntity.history.render();