+++ test_project/test_app/filters.py
@@ -1,18 +1,14 @@
-from django_filters.rest_framework import BooleanFilter
-
 from mapentity.filters import MapEntityFilterSet
 from test_project.test_app.models import DummyModel, Road
 
 
 class DummyModelFilterSet(MapEntityFilterSet):
-    public = BooleanFilter(field_name="public", lookup_expr="exact")
-
-    class Meta:
+    class Meta(MapEntityFilterSet.Meta):
         model = DummyModel
-        fields = ("public", "name")
+        fields = ("public", "name", "tags")
 
 
 class RoadFilterSet(MapEntityFilterSet):
-    class Meta:
+    class Meta(MapEntityFilterSet.Meta):
         model = Road
         fields = ("id", "name")