+++ test_project/test_app/migrations/0020_remove_geopoint_sector_and_more.py
@@ -6,7 +6,7 @@ from django.db import migrations, models
 
 class Migration(migrations.Migration):
     dependencies = [
-        ("test_app", "0018_geopoint"),
+        ("test_app", "0019_road_tag"),
     ]
 
     operations = [