+++ test_project/test_app/tests/test_forms.py
@@ -1,3 +1,4 @@
+from dal import autocomplete
 from django import forms
 from django.test import TestCase
 from django.test.utils import override_settings
@@ -12,6 +13,14 @@ class DummyForm(MapEntityForm):
     class Meta:
         model = DummyModel
         fields = "__all__"
+
+
+class GeopointForm(MapEntityForm):
+    extra_field = forms.CharField()
+
+    class Meta:
+        model = GeoPoint
+        fields = ["located_in", "road", "name", "tags", "extra_field"]
 
 
 class MapEntityFormTest(TestCase):
@@ -39,6 +48,33 @@ class MapEntityFormTest(TestCase):
         form = DummyForm(instance=self.sample_object)
         self.assertIn("name_zh_hant", form.fields)
         self.assertEqual("Name [zh-hant]", form.fields["name_zh_hant"].label)
+
+    def test_m2m_widget(self):
+        # Test that M2M have select2 widget
+        form = GeopointForm()
+        self.assertIn("tags", form.fields)
+        self.assertTrue(
+            isinstance(form.fields["tags"].widget, autocomplete.Select2Multiple)
+        )
+
+    def test_fk_widget(self):
+        # Test that FK have select2 widget
+        form = GeopointForm()
+        self.assertIn("located_in", form.fields)
+        self.assertTrue(
+            isinstance(form.fields["located_in"].widget, autocomplete.ListSelect2)
+        )
+        self.assertIn("road", form.fields)
+        self.assertTrue(
+            isinstance(form.fields["road"].widget, autocomplete.ListSelect2)
+        )
+
+    def test_do_not_change_unwanted_widgets(self):
+        form = GeopointForm()
+        self.assertIn("name_en", form.fields)
+        self.assertTrue(isinstance(form.fields["name_en"].widget, forms.TextInput))
+        self.assertIn("extra_field", form.fields)
+        self.assertTrue(isinstance(form.fields["extra_field"].widget, forms.TextInput))
 
 
 class MapEntityRichTextFormTest(TestCase):